requirements.txt
click==8.1.3
flask==2.2.3
gunicorn==20.1.0
itsdangerous==2.1.2
jinja2==3.1.2
joblib==1.2.0
markupsafe==2.1.2
numpy==1.24.2
pandas==1.5.3
python-dateutil==2.8.2
pytz==2022.7.1
pyyaml==6.0
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
werkzeug==2.2.3